fix responding to channels with no reply subject
This would happen if a PUB was sent on a channel that a client was set
up to respond to. The message would have a blank reply-to channel,
causing a bogus message to be sent in response.

Also: refactor the tests to use gocheck setup/teardown + helpers for
starting/stopping NATS. Also now waits for NATS to actually be down,
potentially eliminating some races.

[finishes #53723169]